+++ sqlmesh/core/test/result.py
@@ -49,7 +49,7 @@ class ModelTextTestResult(unittest.TextTestResult):
             err: A tuple of the form returned by sys.exc_info(), i.e., (type, value, traceback).
         """
         exctype, value, tb = err
-        self.err = (test, err)
+        self.original_err = (test, err)
         return super().addFailure(test, (exctype, value, None))  # type: ignore
 
     def addError(self, test: unittest.TestCase, err: ErrorType) -> None:
@@ -61,7 +61,7 @@ class ModelTextTestResult(unittest.TextTestResult):
             test: The test case.
             err: A tuple of the form returned by sys.exc_info(), i.e., (type, value, traceback).
         """
-        self.err = (test, err)
+        self.original_err = (test, err)
         return super().addError(test, err)  # type: ignore
 
     def addSuccess(self, test: unittest.TestCase) -> None:
